refactor: cancel actions are returns null on android. Also platformName method deleted and tests are updated. Handler and plugin classes seperated to different files.
