+++ document_scanner_kit_android/lib/document_scanner_kit_android.dart
@@ -14,11 +14,6 @@ class DocumentScannerKitAndroid extends DocumentScannerKitPlatform {
   }
 
   @override
-  Future<String?> getPlatformName() {
-    return methodChannel.invokeMethod<String>('getPlatformName');
-  }
-
-  @override
   Future<List<String>?> scan() {
     return methodChannel.invokeListMethod<String>('scan');
   }
