+++ document_scanner_kit_ios/lib/document_scanner_kit_ios.dart
@@ -14,11 +14,6 @@ class DocumentScannerKitIOS extends DocumentScannerKitPlatform {
   }
 
   @override
-  Future<String?> getPlatformName() {
-    return methodChannel.invokeMethod<String>('getPlatformName');
-  }
-
-  @override
   Future<List<String>?> scan() {
     return methodChannel.invokeListMethod<String>('scan');
   }
