+++ document_scanner_kit_ios/test/document_scanner_kit_ios_test.dart
@@ -7,22 +7,18 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   group('DocumentScannerKitIOS', () {
-    const kPlatformName = 'iOS';
     const kScanResult = ['path1', 'path2'];
     late DocumentScannerKitIOS documentScannerKit;
     late List<MethodCall> log;
 
     setUp(() async {
       documentScannerKit = DocumentScannerKitIOS();
-
       log = <MethodCall>[];
       TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
           .setMockMethodCallHandler(documentScannerKit.methodChannel,
               (methodCall) async {
         log.add(methodCall);
         switch (methodCall.method) {
-          case 'getPlatformName':
-            return kPlatformName;
           case 'scan':
             return kScanResult;
           default:
@@ -34,15 +30,6 @@ void main() {
     test('can be registered', () {
       DocumentScannerKitIOS.registerWith();
       expect(DocumentScannerKitPlatform.instance, isA<DocumentScannerKitIOS>());
-    });
-
-    test('getPlatformName returns correct name', () async {
-      final name = await documentScannerKit.getPlatformName();
-      expect(
-        log,
-        <Matcher>[isMethodCall('getPlatformName', arguments: null)],
-      );
-      expect(name, equals(kPlatformName));
     });
 
     test('scan returns correct paths', () async {
